+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 ## [Unreleased]
-*no unreleased changes*
-
+* Handle three char months in Daterange
 ## 5.6.1 / 2020-01-02
 ## Fixed
 * Fix issue with 2020 public holiday